+++ packages/fdc3-agent-proxy/src/DesktopAgentProxy.ts
@@ -55,14 +55,11 @@ export class DesktopAgentProxy implements DesktopAgent, Connectable {
     this.broadcast = this.broadcast.bind(this);
     this.addContextListener = this.addContextListener.bind(this);
     this.getUserChannels = this.getUserChannels.bind(this);
-    this.getSystemChannels = this.getSystemChannels.bind(this);
     this.getOrCreateChannel = this.getOrCreateChannel.bind(this);
     this.createPrivateChannel = this.createPrivateChannel.bind(this);
     this.leaveCurrentChannel = this.leaveCurrentChannel.bind(this);
     this.joinUserChannel = this.joinUserChannel.bind(this);
-    this.joinChannel = this.joinChannel.bind(this);
     this.getCurrentChannel = this.getCurrentChannel.bind(this);
-    this.joinChannel = this.joinChannel.bind(this);
     this.findIntent = this.findIntent.bind(this);
     this.findIntentsByContext = this.findIntentsByContext.bind(this);
     this.raiseIntent = this.raiseIntent.bind(this);
@@ -94,37 +91,17 @@ export class DesktopAgentProxy implements DesktopAgent, Connectable {
     }
   }
 
-  addContextListener(
-    contextTypeOrHandler: ContextHandler | string | null,
-    handler?: ContextHandler
-  ): Promise<Listener> {
-    let theContextType: string | null;
-    let theHandler: ContextHandler;
-
-    if (contextTypeOrHandler == null && typeof handler === 'function') {
-      theContextType = null;
-      theHandler = handler;
-    } else if (typeof contextTypeOrHandler === 'string' && typeof handler === 'function') {
-      theContextType = contextTypeOrHandler;
-      theHandler = handler;
-    } else if (typeof contextTypeOrHandler === 'function') {
-      // deprecated one-arg version
-      theContextType = null;
-      theHandler = contextTypeOrHandler as ContextHandler;
-    } else {
-      //invalid call
-      // TODO: Replace with Standardized error when #1490 is resolved
+  addContextListener(contextType: string | null, handler: ContextHandler): Promise<Listener> {
+    if (typeof contextType !== 'string' && contextType !== null) {
       throw new Error('Invalid arguments passed to addContextListener!');
     }
-
-    return this.channels.addContextListener(theHandler, theContextType);
+    if (typeof handler !== 'function') {
+      throw new Error('Invalid arguments passed to addContextListener!');
+    }
+    return this.channels.addContextListener(handler, contextType);
   }
 
   getUserChannels() {
-    return this.channels.getUserChannels();
-  }
-
-  getSystemChannels() {
     return this.channels.getUserChannels();
   }
 
@@ -144,10 +121,6 @@ export class DesktopAgentProxy implements DesktopAgent, Connectable {
     return this.channels.joinUserChannel(channelId);
   }
 
-  joinChannel(channelId: string) {
-    return this.channels.joinUserChannel(channelId);
-  }
-
   getCurrentChannel(): Promise<Channel | null> {
     return this.channels.getUserChannel();
   }
@@ -160,25 +133,8 @@ export class DesktopAgentProxy implements DesktopAgent, Connectable {
     return this.intents.findIntentsByContext(context);
   }
 
-  private ensureAppId(app?: string | AppIdentifier | null): AppIdentifier | undefined {
-    if (typeof app === 'string') {
-      return {
-        appId: app,
-      };
-    } else if (app?.appId) {
-      return app as AppIdentifier;
-    } else {
-      return undefined;
-    }
-  }
-
-  raiseIntent(
-    intent: string,
-    context: Context,
-    app?: string | AppIdentifier | null,
-    metadata?: AppProvidableContextMetadata
-  ) {
-    return this.intents.raiseIntent(intent, context, this.ensureAppId(app), metadata);
+  raiseIntent(intent: string, context: Context, app?: AppIdentifier | null, metadata?: AppProvidableContextMetadata) {
+    return this.intents.raiseIntent(intent, context, app ?? undefined, metadata);
   }
 
   addIntentListener(intent: string, handler: IntentHandler) {
@@ -191,14 +147,14 @@ export class DesktopAgentProxy implements DesktopAgent, Connectable {
 
   raiseIntentForContext(
     context: Context,
-    app?: string | AppIdentifier | null,
+    app?: AppIdentifier | null,
     metadata?: AppProvidableContextMetadata
   ): Promise<IntentResolution> {
-    return this.intents.raiseIntentForContext(context, this.ensureAppId(app), metadata);
+    return this.intents.raiseIntentForContext(context, app ?? undefined, metadata);
   }
 
-  open(app: string | AppIdentifier, context?: Context | null, metadata?: AppProvidableContextMetadata) {
-    return this.apps.open(this.ensureAppId(app)!, context, metadata);
+  open(app: AppIdentifier, context?: Context | null, metadata?: AppProvidableContextMetadata) {
+    return this.apps.open(app, context, metadata);
   }
 
   close(): Promise<void> {