+++ packages/fdc3-get-agent/test/step-definitions/desktop-agent.steps.ts
@@ -230,8 +230,6 @@ Given('A Dummy Desktop Agent in {string}', async (world: CustomWorld, field: str
     getCurrentChannel: notImplemented,
     leaveCurrentChannel: notImplemented,
     getAppMetadata: notImplemented,
-    getSystemChannels: notImplemented,
-    joinChannel: notImplemented,
   };
 
   world.props[field] = da;