+++ packages/fdc3-security/test/mocks/MockPrivateChannel.ts
@@ -8,15 +8,6 @@ export class MockPrivateChannel extends MockChannel implements PrivateChannel {
   constructor(id: string) {
     super(id, 'private');
   }
-  onAddContextListener(_handler: (contextType?: string) => void): Listener {
-    return { unsubscribe: async () => {} };
-  }
-  onUnsubscribe(_handler: (contextType?: string) => void): Listener {
-    return { unsubscribe: async () => {} };
-  }
-  onDisconnect(_handler: () => void): Listener {
-    return { unsubscribe: async () => {} };
-  }
   async disconnect(): Promise<void> {
     this.disconnectCallCount++;
   }