+++ packages/fdc3-standard/src/api/Channel.ts
@@ -129,10 +129,4 @@ export interface Channel {
    *
    */
   addEventListener(type: string | null, handler: EventHandler): Promise<Listener>;
-
-  /**
-   * @deprecated use `addContextListener(null, handler)` instead of `addContextListener(handler)`.
-   */
-  // eslint-disable-next-line @typescript-eslint/adjacent-overload-signatures
-  addContextListener(handler: ContextHandler): Promise<Listener>;
 }