+++ packages/fdc3-standard/src/api/DesktopAgent.ts
@@ -376,17 +376,21 @@ export interface DesktopAgent {
    *   const symbol = context.id.symbol;
    *
    *   // Called when the remote side adds a context listener
-   *   const addContextListener = channel.onAddContextListener((contextType) => {
-   *     // broadcast price quotes as they come in from our quote feed
-   *     feed.onQuote(symbol, (price) => {
-   *       channel.broadcast({ type: "price", price});
-   *     });
-   *   });
+   *   const addContextListener = await channel.addEventListener("addContextListener",
+   *     (event) => {
+   *       // broadcast price quotes as they come in from our quote feed
+   *       feed.onQuote(symbol, (price) => {
+   *         channel.broadcast({ type: "price", price});
+   *       });
+   *     }
+   *   );
    *
    *   // Stop the feed if the remote side closes
-   *   const disconnectListener = channel.onDisconnect(() => {
-   *     feed.stop(symbol);
-   *   });
+   *   const disconnectListener = await channel.addEventListener("disconnect",
+   *     () => {
+   *       feed.stop(symbol);
+   *     }
+   *   );
    *
    *   return channel;
    * });
@@ -532,22 +536,28 @@ export interface DesktopAgent {
    * 	const symbol = context.id.ticker;
    *
    * 	// This gets called when the remote side adds a context listener
-   * 	const addContextListener = channel.onAddContextListener((contextType) => {
-   * 		// broadcast price quotes as they come in from our quote feed
-   * 		feed.onQuote(symbol, (price) => {
-   * 			channel.broadcast({ type: "price", price});
-   * 		});
-   * 	});
+   * 	const addContextListener = await channel.addEventListener("addContextListener",
+   * 		(event) => {
+   * 			// broadcast price quotes as they come in from our quote feed
+   * 			feed.onQuote(symbol, (price) => {
+   * 				channel.broadcast({ type: "price", price});
+   * 			});
+   * 		}
+   * 	);
    *
    * 	// This gets called when the remote side calls Listener.unsubscribe()
-   * 	const unsubscriberListener = channel.onUnsubscribe((contextType) => {
-   * 		feed.stop(symbol);
-   * 	});
+   * 	const unsubscribeListener = await channel.addEventListener("unsubscribe",
+   * 		(event) => {
+   * 			feed.stop(symbol);
+   * 		}
+   * 	);
    *
    * 	// This gets called if the remote side closes
-   * 	const disconnectListener = channel.onDisconnect(() => {
-   * 		feed.stop(symbol);
-   * 	})
+   * 	const disconnectListener = await channel.addEventListener("disconnect",
+   * 		() => {
+   * 			feed.stop(symbol);
+   * 		}
+   * 	);
    *
    * 	return channel;
    * });
@@ -620,58 +630,4 @@ export interface DesktopAgent {
    * ```
    */
   getAppMetadata(app: AppIdentifier): Promise<AppMetadata>;
-
-  //---------------------------------------------------------------------------------------------
-  // Deprecated function signatures
-  //---------------------------------------------------------------------------------------------
-
-  /**
-   * Adds a listener for incoming context broadcasts from the Desktop Agent.
-   * @deprecated use `addContextListener(null, handler)` instead of `addContextListener(handler)`. Provided for backwards compatibility with versions FDC3 standard <2.0.
-   */
-  // eslint-disable-next-line @typescript-eslint/adjacent-overload-signatures
-  addContextListener(handler: ContextHandler): Promise<Listener>;
-
-  /**
-   * @deprecated Alias to the `getUserChannels` function provided for backwards compatibility with versions FDC3 standard <2.0.
-   */
-  getSystemChannels(): Promise<Array<Channel>>;
-
-  /**
-   * @deprecated Alias to the `joinUserChannel` function Provided for backwards compatibility with versions FDC3 standard <2.0.
-   */
-  joinChannel(channelId: string): Promise<void>;
-
-  /**
-   * @deprecated version of `open` that launches an app by by name rather than `AppIdentifier`. Provided for backwards compatibility with versions FDC3 standard <2.0.
-   *
-   * ```javascript
-   * //Open an app without context, using the app name
-   * let instanceMetadata = await fdc3.open('myApp');
-   * ```
-   */
-  // eslint-disable-next-line @typescript-eslint/adjacent-overload-signatures
-  open(name: string, context?: Context): Promise<AppIdentifier>;
-
-  /**
-   * @deprecated version of `raiseIntent` that targets an app by by name rather than `AppIdentifier`. Provided for backwards compatibility with versions FDC3 standard <2.0.
-   *
-   * ```javascript
-   * // use the `name` metadata of an app to describe the target app for the intent
-   * await fdc3.raiseIntent("StartChat", context, appIntent.apps[0].name);
-   * ```
-   */
-  // eslint-disable-next-line @typescript-eslint/adjacent-overload-signatures
-  raiseIntent(intent: Intent, context: Context, name: string): Promise<IntentResolution>;
-
-  /**
-   * @deprecated version of `raiseIntentForContext` that targets an app by by name rather than `AppIdentifier`. Provided for backwards compatibility with versions FDC3 standard <2.0.
-   *
-   * ```javascript
-   * // Resolve against all intents registered by a specific target app name for the specified context
-   * await fdc3.raiseIntentForContext(context, targetAppName);
-   * ```
-   */
-  // eslint-disable-next-line @typescript-eslint/adjacent-overload-signatures
-  raiseIntentForContext(context: Context, name: string): Promise<IntentResolution>;
 }