+++ packages/fdc3-standard/src/api/IntentMetadata.ts
@@ -13,7 +13,7 @@ export interface IntentMetadata {
   readonly name: Intent;
 
   /** Display name for the intent.
-   * @deprecated Use the intent name for display as display name may vary for
+   * @deprecated Since FDC3 2.1. Use the intent name for display as display name may vary for
    * each application as it is defined in the app's AppD record.
    */
   readonly displayName?: string;